refactor(desktop): integrate tauri-specta and restructure types
- Integrated `tauri-specta` to auto-generate type-safe TypeScript bindings for Tauri commands.
- Replaced manual `invoke` strings with generated `commands` in `tauriApi.ts`.
- Split `types/index.ts` into modular files: `prompt.ts`, `review.ts`, and `settings.ts`.
- Removed manual type definitions that are now generated in `bindings.ts`.
- Updated Rust structs and commands with `#[derive(Type)]` and `#[specta::specta]` macros.
- Added `biome.json` configuration and updated project scripts for linting/checking.
- Refactored Rust tests and extracted magic prompt defaults.

+++ apps/desktop/src-tauri/src/commands.rs
@@ -16,6 +16,7 @@ use tauri::Manager;
 use uuid::Uuid;
 
 #[tauri::command]
+#[specta::specta]
 pub fn open_project_window(app: tauri::AppHandle, root_path: String) -> Result<()> {
     debug!("Attempting to open project window for path: {root_path}");
 
@@ -73,6 +74,7 @@ pub fn open_project_window(app: tauri::AppHandle, root_path: String) -> Result<(
 }
 
 #[tauri::command]
+#[specta::specta]
 pub fn register_window_project(window: tauri::Window, root_path: Option<String>) -> Result<()> {
     debug!(
         "Registering project for window '{}': {:?}",
@@ -89,6 +91,7 @@ pub fn register_window_project(window: tauri::Window, root_path: Option<String>)
 }
 
 #[tauri::command]
+#[specta::specta]
 pub fn create_new_window(app: tauri::AppHandle) -> Result<()> {
     let label = format!("main-{}", Uuid::new_v4());
     debug!("Creating new blank window with label: {label}");
@@ -104,6 +107,7 @@ pub fn create_new_window(app: tauri::AppHandle) -> Result<()> {
 }
 
 #[tauri::command]
+#[specta::specta]
 pub async fn close_window(window: tauri::Window) -> Result<()> {
     debug!(
         "Closing window '{}' and unregistering project.",
@@ -120,11 +124,13 @@ pub async fn close_window(window: tauri::Window) -> Result<()> {
 }
 
 #[tauri::command]
+#[specta::specta]
 pub async fn list_directory_recursive(path: String, settings: IgnoreSettings) -> Result<FileNode> {
     Ok(project_service::list_directory_recursive(&PathBuf::from(path), settings).await?)
 }
 
 #[tauri::command]
+#[specta::specta]
 pub fn get_relative_path(full_path: String, root_path: String) -> Result<String> {
     Ok(path_utils::get_relative_path(
         &PathBuf::from(full_path),
@@ -133,45 +139,53 @@ pub fn get_relative_path(full_path: String, root_path: String) -> Result<String>
 }
 
 #[tauri::command]
+#[specta::specta]
 pub async fn read_file_as_base64(path: String) -> Result<String> {
     let bytes = project_service::read_file_bytes(&PathBuf::from(path)).await?;
     Ok(general_purpose::STANDARD.encode(bytes))
 }
 
 #[tauri::command]
+#[specta::specta]
 pub async fn read_file_content(path: String) -> Result<String> {
     Ok(project_service::read_file_content(&PathBuf::from(path)).await?)
 }
 
 #[tauri::command]
+#[specta::specta]
 pub async fn is_binary_file(path: String) -> Result<bool> {
     Ok(project_service::is_binary(&PathBuf::from(path)).await?)
 }
 
 #[tauri::command]
+#[specta::specta]
 pub async fn file_exists(path: String) -> Result<bool> {
     Ok(PathBuf::from(path).exists())
 }
 
 #[tauri::command]
+#[specta::specta]
 pub async fn write_file_content(path: String, content: String) -> Result<()> {
     project_service::write_file_content(&PathBuf::from(path), &content).await?;
     Ok(())
 }
 
 #[tauri::command]
+#[specta::specta]
 pub async fn delete_file(file_path: String) -> Result<()> {
     project_service::delete_file(&PathBuf::from(file_path)).await?;
     Ok(())
 }
 
 #[tauri::command]
+#[specta::specta]
 pub async fn move_file(from: String, to: String) -> Result<()> {
     project_service::move_file(&PathBuf::from(from), &PathBuf::from(to)).await?;
     Ok(())
 }
 
 #[tauri::command]
+#[specta::specta]
 pub async fn backup_files(root_path: String, file_paths: Vec<String>) -> Result<String> {
     let root = PathBuf::from(root_path);
     let paths = file_paths.into_iter().map(PathBuf::from).collect();
@@ -179,6 +193,7 @@ pub async fn backup_files(root_path: String, file_paths: Vec<String>) -> Result<
 }
 
 #[tauri::command]
+#[specta::specta]
 pub async fn revert_file_from_backup(
     root_path: String,
     backup_id: String,
@@ -194,17 +209,20 @@ pub async fn revert_file_from_backup(
 }
 
 #[tauri::command]
+#[specta::specta]
 pub async fn read_file_from_backup(backup_id: String, relative_path: String) -> Result<String> {
     Ok(review_service::read_file_from_backup(&backup_id, &PathBuf::from(relative_path)).await?)
 }
 
 #[tauri::command]
+#[specta::specta]
 pub async fn delete_backup(backup_id: String) -> Result<()> {
     review_service::delete_backup(&backup_id).await?;
     Ok(())
 }
 
 #[tauri::command]
+#[specta::specta]
 pub async fn parse_changes_from_markdown(
     markdown: String,
     root_path: String,
@@ -213,16 +231,19 @@ pub async fn parse_changes_from_markdown(
 }
 
 #[tauri::command]
+#[specta::specta]
 pub async fn is_git_repository(path: String) -> Result<bool> {
     Ok(git_service::is_git_repository(&PathBuf::from(path))?)
 }
 
 #[tauri::command]
+#[specta::specta]
 pub async fn get_git_status(repo_path: String) -> Result<GitStatus> {
     Ok(git_service::get_git_status(&PathBuf::from(repo_path))?)
 }
 
 #[tauri::command]
+#[specta::specta]
 pub async fn get_recent_commits(repo_path: String, count: u32) -> Result<Vec<Commit>> {
     Ok(git_service::get_recent_commits(
         &PathBuf::from(repo_path),
@@ -231,6 +252,7 @@ pub async fn get_recent_commits(repo_path: String, count: u32) -> Result<Vec<Com
 }
 
 #[tauri::command]
+#[specta::specta]
 pub async fn get_git_diff(repo_path: String, option: DiffOption) -> Result<String> {
     Ok(git_service::get_git_diff(
         &PathBuf::from(repo_path),
@@ -239,11 +261,13 @@ pub async fn get_git_diff(repo_path: String, option: DiffOption) -> Result<Strin
 }
 
 #[tauri::command]
+#[specta::specta]
 pub async fn resolve_path(path: String, cwd: Option<String>) -> Result<String> {
     Ok(path_utils::resolve_path(&path, cwd)?)
 }
 
 #[tauri::command]
+#[specta::specta]
 pub async fn start_pty_session(
     root_path: String,
     command: Option<String>,
@@ -254,34 +278,40 @@ pub async fn start_pty_session(
 }
 
 #[tauri::command]
+#[specta::specta]
 pub async fn resize_pty(rows: u16, cols: u16) -> Result<()> {
     pty_service::resize_pty(rows, cols)?;
     Ok(())
 }
 
 #[tauri::command]
+#[specta::specta]
 pub async fn write_to_pty(text: String) -> Result<()> {
     pty_service::write_to_pty(text).await?;
     Ok(())
 }
 
 #[tauri::command]
+#[specta::specta]
 pub async fn kill_pty() -> Result<()> {
     pty_service::kill_pty()?;
     Ok(())
 }
 
 #[tauri::command]
+#[specta::specta]
 pub async fn get_cli_status() -> Result<CliStatusResult> {
     Ok(cli_service::get_cli_status())
 }
 
 #[tauri::command]
+#[specta::specta]
 pub async fn install_cli_shim() -> Result<CliInstallResult> {
     Ok(cli_service::install_cli_shim().await?)
 }
 
 #[tauri::command]
+#[specta::specta]
 pub async fn start_watching(
     app_handle: tauri::AppHandle,
     root_path: String,
@@ -292,17 +322,19 @@ pub async fn start_watching(
 }
 
 #[tauri::command]
+#[specta::specta]
 pub async fn stop_watching(root_path: String) -> Result<()> {
     watcher_service::stop_watching(&PathBuf::from(root_path));
     Ok(())
 }
 
 #[tauri::command]
+#[specta::specta]
 pub async fn search_files(
     query: String,
     root_path: String,
     settings: IgnoreSettings,
-    limit: Option<usize>,
+    limit: Option<u32>,
 ) -> Result<Vec<SearchResult>> {
     Ok(
         file_search_service::search_files(&PathBuf::from(root_path), &query, settings, limit)
